refactor: 메타데이터 관련 유틸리티 파일명 변경 및 개선
